See Changlogs for details of Sprint 3

+++ app/Http/Controllers/AdminUsersController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\AdminUsers;
+use http\Client\Response;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class AdminUsersController extends Controller
@@ -10,15 +12,47 @@ class AdminUsersController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return array
+     * @return JsonResponse
      */
-    public function index() : array
+    /**
+     * @OA\Get(
+     *      path="/api/adminUsers",
+     *      operationId="indexAdminUsers",
+     *      tags={"AdminUsers"},
+     *      summary="Get list of AdminUsers",
+     *      description="Returns list of AdminUsers",
+     *      security={{ "bearer_token": {} }},
+     *      @OA\Parameter(
+     *         name="idUser",
+     *         in="query",
+     *         description="id of the User who ask the request",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation",
+     *           @OA\JsonContent(
+     *             @OA\Property(property="status", type="integer", example="200"),
+     *             @OA\Property(property="data",type="object"),
+     *             @OA\Property(property="message",type="string")
+     *          )
+     *       )
+     * )
+     *
+     * Returns list of AdminUsers
+     */
+    public function index(Request $request) : JsonResponse
     {
-        $user = AdminUsers::latest()->paginate(10);
-        return [
-            "status" => 1,
-            "data" => $user
-        ];
+        $request->validate([
+            'idUser' => 'required'
+        ]);
+        if($this->isAdminUser($request->get('idUser'))) {
+            $user = AdminUsers::all();
+            return $this->sendResponse($user, "Authorised to see list of Admin users");
+        }
+        else {
+            return $this->sendError("Request not found");
+        }
     }
 
     /**
@@ -35,33 +69,92 @@ class AdminUsersController extends Controller
      * Store a newly created resource in storage.
      *
      * @param Request $request
-     * @return array
+     * @return JsonResponse
      */
-    public function store(Request $request) : array
+    /**
+     * @OA\Post(
+     *      path="/api/adminUsers",
+     *      operationId="storeAdminUsers",
+     *      tags={"AdminUsers"},
+     *      summary="Grant an Users",
+     *      description="Returns a AdminUsers",
+     *      security={{ "bearer_token": {} }},
+     *      @OA\Parameter(
+     *         name="idUser",
+     *         in="query",
+     *         description="id of the User to be granted",
+     *         required=true,
+     *      ),
+     *      @OA\Parameter(
+     *         name="grantedBy",
+     *         in="query",
+     *         description="id of the AdminUser",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation",
+     *           @OA\JsonContent(
+     *             @OA\Property(property="status", type="integer", example="200"),
+     *             @OA\Property(property="data",type="object"),
+     *             @OA\Property(property="message",type="string")
+     *          )
+     *       )
+     * )
+     *
+     * Returns a AdminUsers
+     */
+    public function store(Request $request) : JsonResponse
     {
         $request->validate([
             'idUser' => 'required',
             'grantedBy' => 'required',
         ]);
-        $user = AdminUsers::create($request->all());
-        return [
-            'status' => 1,
-            "data" => $user,
-        ];
+        if($this->isAdminUser($request->get('grantedBy'))) {
+            $user = AdminUsers::create($request->all());
+            return $this->sendResponse($user, "User granted successfully");
+        }
+        else{
+            return $this->sendError("Request not found");
+        }
     }
 
     /**
      * Display the specified resource.
      *
      * @param AdminUsers $user
-     * @return array
+     * @return JsonResponse
      */
-    public function show(AdminUsers $user) : array
+    /**
+     * @OA\Get(
+     *      path="/api/adminUsers/{idUser}",
+     *      operationId="showAdminUsers",
+     *      tags={"AdminUsers"},
+     *      summary="Get the AdminUser",
+     *      description="Returns the AdminUser",
+     *      security={{ "bearer_token": {} }},
+     *      @OA\Parameter(
+     *         name="idUser",
+     *         in="path",
+     *         description="id of the User",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation",
+     *           @OA\JsonContent(
+     *             @OA\Property(property="status", type="integer", example="200"),
+     *             @OA\Property(property="data",type="object"),
+     *             @OA\Property(property="message",type="string")
+     *          )
+     *       )
+     * )
+     *
+     * Returns the AdminUser
+     */
+    public function show(AdminUsers $user) : JsonResponse
     {
-        return [
-            "status" => 1,
-            "data" => $user
-        ];
+        return $this->sendResponse($user, "Result");
     }
 
     /**
@@ -80,28 +173,54 @@ class AdminUsersController extends Controller
      *
      * @param Request $request
      * @param AdminUsers $user
-     * @return array
+     * @return JsonResponse
      */
-    public function update(Request $request, AdminUsers $user) : array
+    /**
+     * @OA\Patch(
+     *      path="/api/adminUsers/{idUser}",
+     *      operationId="updateAdminUsers",
+     *      tags={"AdminUsers"},
+     *      summary="Update the AdminUser",
+     *      description="Returns the AdminUser",
+     *      security={{ "bearer_token": {} }},
+     *      @OA\Parameter(
+     *         name="idUser",
+     *         in="path",
+     *         description="id of the User",
+     *         required=true,
+     *      ),
+     *      @OA\Parameter(
+     *         name="grantedBy",
+     *         in="query",
+     *         description="id of the AdminUser who ask the request",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation",
+     *           @OA\JsonContent(
+     *             @OA\Property(property="status", type="integer", example="200"),
+     *             @OA\Property(property="data",type="object"),
+     *             @OA\Property(property="message",type="string")
+     *          )
+     *       )
+     * )
+     *
+     * Returns the AdminUser
+     */
+    public function update(Request $request, AdminUsers $user) : JsonResponse
     {
         $request->validate([
-            'idUser' => 'required',
             'grantedBy' => 'required',
         ]);
-        if(AdminUsers::where('idUser', $request->get('grantedBy')) != null)
+        if($this->isAdminUser($request->get('grantedBy')))
         {
             $user->update($request->all());
-            return [
-                "status" => 1,
-                "data" => $user,
-                "msg" => "Rights updated successfully"
-            ];
+            return $this->sendResponse($user, "Rights updated successfully");
         }
         else
         {
-            return [
-                "status" => 500,
-            ];
+            return $this->sendError("Request not found");
         }
     }
 
@@ -109,15 +228,54 @@ class AdminUsersController extends Controller
      * Remove the specified resource from storage.
      *
      * @param AdminUsers $user
-     * @return array
+     * @return JsonResponse
      */
-    public function destroy(AdminUsers $user): array
+    /**
+     * @OA\Delete(
+     *      path="/api/adminUsers/{idUser}",
+     *      operationId="deleteAdminUsers",
+     *      tags={"AdminUsers"},
+     *      summary="Delete the AdminUser",
+     *      description="Returns the AdminUser",
+     *      security={{ "bearer_token": {} }},
+     *      @OA\Parameter(
+     *         name="idUser",
+     *         in="path",
+     *         description="id of the User",
+     *         required=true,
+     *      ),
+     *      @OA\Parameter(
+     *         name="grantedBy",
+     *         in="query",
+     *         description="id of the AdminUser who ask the request",
+     *         required=true,
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation",
+     *           @OA\JsonContent(
+     *             @OA\Property(property="status", type="integer", example="200"),
+     *             @OA\Property(property="data",type="object"),
+     *             @OA\Property(property="message",type="string")
+     *          )
+     *       )
+     * )
+     *
+     * Returns the AdminUser
+     */
+    public function destroy(Request $request, AdminUsers $user): JsonResponse
     {
-        $user->delete();
-        return [
-            "status" => 1,
-            "data" => $user,
-            "msg" => "Rights deleted successfully"
-        ];
+        $request->validate([
+            'grantedBy' => 'required',
+        ]);
+        if($this->isAdminUser($request->get('grantedBy')))
+        {
+            $user->delete();
+            return $this->sendResponse($user, "Rights deleted successfully");
+        }
+        else
+        {
+            return $this->sendError("Request not found");
+        }
     }
 }
